requirements.txt
aiogram
python-dotenv
sqlalchemy
aiosqlite
greenlet